requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
attrs==23.2.0
Brotli==1.1.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
discord.py==2.3.2
frozenlist==1.4.1
idna==3.6
multidict==6.0.5
mutagen==1.47.0
pycparser==2.22
pycryptodomex==3.20.0
PyNaCl==1.5.0
python-dotenv==1.0.1
requests==2.31.0
setuptools==69.2.0
typing_extensions==4.10.0
urllib3==2.2.1
websockets==12.0
yarl==1.9.4
youtube-dl==2021.12.17
yt-dlp @ https://github.com/yt-dlp/yt-dlp/archive/master.tar.gz#sha256=db8cb450b239b4ee8692d8d69867670ca3984aecfac0867d25563fd7a604d46a
